+++ test/test_charm.py
@@ -431,10 +431,10 @@ foo-bar:
   title: foo-bar
 start:
   description: "Start the unit."
+  additionalProperties: false
 ''')
 
     def _setup_test_action(self):
-        os.environ['JUJU_ACTION_NAME'] = 'foo-bar'
         fake_script(self, 'action-get', """echo '{"foo-name": "name", "silent": true}'""")
         fake_script(self, 'action-set', "")
         fake_script(self, 'action-log', "")
@@ -475,11 +475,6 @@ start:
             ['action-set', "res=val with spaces"],
             ['action-fail', "test-fail"],
         ])
-
-        # Make sure that action events that do not match the current context are
-        # not possible to emit by hand.
-        with self.assertRaises(RuntimeError):
-            charm.on.start_action.emit()
 
     def test_invalid_action_results(self):