+++ CHANGES.md
@@ -3,6 +3,7 @@
 * Added `Unit.reboot()` and `Harness.reboot_count``
 * Added `RelationMeta.optional`
 * The type of a `Handle`'s `key` was expanded from `str` to `str|None`
+* Added `Harness.run_action()`, `testing.ActionOutput`, and `testing.ActionFailed`
 * Narrowed types of `app` and `unit` in relation events to exclude `None` where applicable
 
 # 2.7.0